Use the new structure to factor out creation of test schemas
