+++ modules/tests/src/main/scala/Person.scala
@@ -1,8 +1,10 @@
-package scalaz.schema
+package scalaz
 
-import scalaz._
-import monocle.{ Getter, Iso }
-import monocle.macros.GenPrism
+package schema
+
+package tests
+
+import monocle.Iso
 
 final case class Person(name: String, role: Option[Role])
 sealed trait Role
@@ -10,40 +12,50 @@ final case class User(active: Boolean)       extends Role
 final case class Admin(rights: List[String]) extends Role
 
 object Person {
-  val name = Getter[Person, String](_.name)
-
-  def setRole(r1: Role): Person => Person =
-    (p: Person) => p.copy(role = Some(r1))
-  val role = monocle.Optional[Person, Role](_.role)(setRole(_))
-
-  val roleEither: Role => User \/ Admin = r =>
-    r match {
-      case u: User  => -\/(u)
-      case a: Admin => \/-(a)
-    }
-
-  val eitherRole: User \/ Admin => Role = e =>
-    e match {
-      case -\/(x) => x
-      case \/-(x) => x
-    }
-
-  val userActive: User => Boolean        = u => u.active
-  val adminRights: Admin => List[String] = a => a.rights
-
-  val personProduct: Person => (String, Option[Role]) = p => (p.name, p.role)
-
-  val productPerson: ((String, Option[Role])) => Person = {
-    case (name, role) => Person(name, role)
-  }
-
-  val user   = GenPrism[Role, User]
-  val admin  = GenPrism[Role, Admin]
-  val active = Getter[User, Boolean](_.active)
-  val rights = Getter[Admin, List[String]](_.rights)
 
   private def f(p: Person): (Seq[Char], Option[Role]) = (p.name.toSeq, p.role)
   private def g(t: (Seq[Char], Option[Role])): Person = Person(t._1.mkString, t._2)
 
   val personToTupleIso = Iso[Person, (Seq[Char], Option[Role])](f)(g)
+}
+
+trait TestModule extends SchemaModule[JsonSchema.type] {
+  val R = JsonSchema
+
+  type PersonTuple = (Seq[Char], Option[Role])
+
+  val user = record(
+    "active" -*>: prim(JsonSchema.JsonBool),
+    Iso[Boolean, User](User.apply)(_.active)
+  )
+
+  val admin = record(
+    "rights" -*>: seq(prim(JsonSchema.JsonString)),
+    Iso[List[String], Admin](Admin.apply)(_.rights)
+  )
+
+  val role = union(
+    "user" -+>: user :+:
+      "admin" -+>: admin,
+    Iso[User \/ Admin, Role] {
+      case -\/(u) => u
+      case \/-(a) => a
+    } {
+      case u @ User(_)  => -\/(u)
+      case a @ Admin(_) => \/-(a)
+    }
+  )
+
+  val person = record(
+    "name" -*>: prim(JsonSchema.JsonString) :*:
+      "role" -*>: optional(
+      role
+    ),
+    Iso[(String, Option[Role]), Person]((Person.apply _).tupled)(p => (p.name, p.role))
+  )
+
+  val personTupleSchema = iso[Person, PersonTuple](
+    person,
+    Person.personToTupleIso
+  )
 }
